+++ app/Http/ViewComposers/IngameMainComposer.php
@@ -122,8 +122,8 @@ class IngameMainComposer
         $view->with([
             'underAttack' => $this->fleetMissionService->currentPlayerUnderAttack(),
             'unreadMessagesCount' => $this->messageService->getUnreadMessagesCount(),
-            'buddyRequestCount' => $this->buddyService->getUnreadRequestsCount(auth()->id()),
-            'onlineBuddiesCount' => $this->buddyService->getOnlineBuddiesCount(auth()->id()),
+            'buddyRequestCount' => $this->buddyService->getUnreadRequestsCount((int) auth()->id()),
+            'onlineBuddiesCount' => $this->buddyService->getOnlineBuddiesCount((int) auth()->id()),
             'resources' => $resources,
             'currentPlayer' => $this->player,
             'currentPlanet' => $this->player->planets->current(),
